moved more components from Dashboard
