+++ src/pages/workspaces/workspace/Dashboard/Dashboard.test.ts
@@ -1,333 +1,11 @@
-import { asMockedFn } from '@terra-ui-packages/test-utils';
-import { act, screen } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { screen } from '@testing-library/react';
 import { axe } from 'jest-axe';
 import _ from 'lodash/fp';
 import { h } from 'react-hyperscript-helpers';
-import { Ajax } from 'src/libs/ajax';
 import { azureRegions } from 'src/libs/azure-regions';
-import { authStore } from 'src/libs/state';
 import { AzureStorageDetails } from 'src/pages/workspaces/workspace/Dashboard/AzureStorageDetails';
-import { BucketLocation } from 'src/pages/workspaces/workspace/Dashboard/BucketLocation';
-import { WorkspaceNotifications } from 'src/pages/workspaces/workspace/Dashboard/Dashboard';
-import { WorkspaceInformation } from 'src/pages/workspaces/workspace/Dashboard/WorkspaceInformation';
 import { renderWithAppContexts as render } from 'src/testing/test-utils';
-import {
-  defaultAzureStorageOptions,
-  defaultGoogleBucketOptions,
-  defaultGoogleWorkspace,
-  protectedAzureWorkspace,
-  regionRestrictedAzureWorkspace,
-} from 'src/testing/workspace-fixtures';
-
-type AjaxContract = ReturnType<typeof Ajax>;
-
-jest.mock('src/libs/ajax');
-
-jest.mock('src/libs/notifications');
-
-describe('WorkspaceInformation', () => {
-  it('renders information for a non-protected workspace without region constraints and does not fail accessibility tests', async () => {
-    // Act
-    const { container } = render(
-      h(WorkspaceInformation, { workspace: { ...defaultGoogleWorkspace, workspaceInitialized: true } })
-    );
-
-    // Assert
-    // Access Level
-    expect(screen.getAllByText('Owner')).not.toBeNull();
-    // Created date
-    expect(screen.getAllByText('2/15/2023')).not.toBeNull();
-    // Last updated date
-    expect(screen.getAllByText('3/15/2023')).not.toBeNull();
-    // Should not have workspace protected entry
-    expect(screen.queryByText('Workspace Protected')).toBeNull();
-    // Should not have region constraint
-    expect(screen.queryByText('Region Constraint')).toBeNull();
-
-    // Accessibility
-    expect(await axe(container)).toHaveNoViolations();
-  });
-
-  it('renders information for a protected workspace and does not fail accessibility tests', async () => {
-    const user = userEvent.setup();
-
-    // Act
-    const { container } = render(
-      h(WorkspaceInformation, { workspace: { ...protectedAzureWorkspace, workspaceInitialized: true } })
-    );
-
-    // Assert
-    // Access Level
-    expect(screen.getAllByText('Owner')).not.toBeNull();
-    // Created date
-    expect(screen.getAllByText('2/15/2023')).not.toBeNull();
-    // Last updated date
-    expect(screen.getAllByText('3/15/2023')).not.toBeNull();
-    // Should show protected workspace information.
-    expect(screen.getAllByText('Workspace Protected')).not.toBeNull();
-    // Should not have region constraint
-    expect(screen.queryByText('Region Constraint')).toBeNull();
-
-    // Act, click on the info button to get tooltip text to render.
-    await user.click(screen.getByLabelText('More info'));
-
-    // Assert
-    expect(screen.getAllByText(/protected or sensitive data/)).not.toBeNull();
-
-    // Accessibility
-    expect(await axe(container)).toHaveNoViolations();
-  });
-
-  it('renders information for a workspace with region constraints and does not fail accessibility tests', async () => {
-    const user = userEvent.setup();
-
-    // Act
-    const { container } = render(
-      h(WorkspaceInformation, { workspace: { ...regionRestrictedAzureWorkspace, workspaceInitialized: true } })
-    );
-
-    // Assert
-    // Access Level
-    expect(screen.getAllByText('Owner')).not.toBeNull();
-    // Created date
-    expect(screen.getAllByText('2/15/2023')).not.toBeNull();
-    // Last updated date
-    expect(screen.getAllByText('3/15/2023')).not.toBeNull();
-    // Should not have workspace protected entry
-    expect(screen.queryByText('Workspace Protected')).toBeNull();
-    // Should show region constraint information.
-    expect(screen.getAllByText('Region Constraint')).not.toBeNull();
-
-    // Act, click on the info button to get tooltip text with region labels to render.
-    await user.click(screen.getByLabelText('More info'));
-
-    // Assert
-    expect(screen.getAllByText(new RegExp(`${azureRegions.eastus.label}`))).not.toBeNull();
-    expect(screen.getAllByText(new RegExp(`${azureRegions.westus2.label}`))).not.toBeNull();
-    expect(screen.getAllByText(/unknownRegion/)).not.toBeNull();
-
-    // Accessibility
-    expect(await axe(container)).toHaveNoViolations();
-  });
-});
-
-describe('WorkspaceNotifications', () => {
-  const testWorkspace = { workspace: { namespace: 'test', name: 'test' } };
-
-  afterEach(() => {
-    authStore.reset();
-    jest.resetAllMocks();
-  });
-
-  it.each([
-    {
-      profile: {
-        'notifications/SuccessfulSubmissionNotification/test/test': 'true',
-        'notifications/FailedSubmissionNotification/test/test': 'true',
-        'notifications/AbortedSubmissionNotification/test/test': 'true',
-      },
-      expectedState: true,
-    },
-    {
-      profile: {},
-      expectedState: true,
-    },
-    {
-      profile: {
-        'notifications/SuccessfulSubmissionNotification/test/test': 'false',
-        'notifications/FailedSubmissionNotification/test/test': 'false',
-        'notifications/AbortedSubmissionNotification/test/test': 'false',
-      },
-      expectedState: false,
-    },
-  ])('renders checkbox with submission notifications status', ({ profile, expectedState }) => {
-    // @ts-expect-error
-    authStore.update((state) => ({ ...state, profile }));
-
-    const { getByLabelText } = render(h(WorkspaceNotifications, { workspace: testWorkspace }));
-    const submissionNotificationsCheckbox = getByLabelText('Receive submission notifications');
-    expect(submissionNotificationsCheckbox.getAttribute('aria-checked')).toBe(`${expectedState}`);
-  });
-
-  it('updates preferences when checkbox is clicked', async () => {
-    const user = userEvent.setup();
-
-    const setPreferences = jest.fn().mockReturnValue(Promise.resolve());
-    asMockedFn(Ajax).mockImplementation(
-      () =>
-        ({
-          Metrics: {
-            captureEvent: jest.fn(),
-          } as Partial<AjaxContract['Metrics']>,
-          User: {
-            profile: {
-              get: jest.fn().mockReturnValue(Promise.resolve({ keyValuePairs: [] })),
-              setPreferences,
-            } as Partial<AjaxContract['User']['profile']>,
-          } as Partial<AjaxContract['User']>,
-        } as Partial<AjaxContract> as AjaxContract)
-    );
-
-    authStore.update((state) => ({
-      ...state,
-      profile: {
-        // @ts-expect-error
-        'notifications/SuccessfulSubmissionNotification/test/test': 'false',
-        'notifications/FailedSubmissionNotification/test/test': 'false',
-        'notifications/AbortedSubmissionNotification/test/test': 'false',
-      },
-    }));
-
-    const { getByLabelText } = render(h(WorkspaceNotifications, { workspace: testWorkspace }));
-    const submissionNotificationsCheckbox = getByLabelText('Receive submission notifications');
-
-    await user.click(submissionNotificationsCheckbox);
-    expect(setPreferences).toHaveBeenCalledWith({
-      'notifications/SuccessfulSubmissionNotification/test/test': 'true',
-      'notifications/FailedSubmissionNotification/test/test': 'true',
-      'notifications/AbortedSubmissionNotification/test/test': 'true',
-    });
-  });
-
-  it('has no accessibility errors', async () => {
-    authStore.update((state) => ({
-      ...state,
-      profile: {
-        // @ts-expect-error
-        'notifications/SuccessfulSubmissionNotification/test/test': 'false',
-        'notifications/FailedSubmissionNotification/test/test': 'false',
-        'notifications/AbortedSubmissionNotification/test/test': 'false',
-      },
-    }));
-
-    const { container } = render(h(WorkspaceNotifications, { workspace: testWorkspace }));
-    expect(await axe(container)).toHaveNoViolations();
-  });
-});
-
-describe('BucketLocation', () => {
-  const workspace = {
-    workspace: { namespace: 'test', name: 'test', cloudPlatform: 'Gcp' },
-    workspaceInitialized: true,
-  };
-
-  afterEach(() => {
-    jest.resetAllMocks();
-  });
-
-  it('shows Loading initially when uninitialized and should not fail any accessibility tests', async () => {
-    // Arrange
-    const props = {
-      workspace: _.merge(workspace, { workspaceInitialized: false }),
-      storageDetails: _.merge(defaultGoogleBucketOptions, defaultAzureStorageOptions),
-    };
-
-    // Act
-    const { container } = render(h(BucketLocation, props));
-
-    // Assert
-    expect(screen.queryByText('Loading')).not.toBeNull();
-    expect(await axe(container)).toHaveNoViolations();
-  });
-
-  it('shows Loading initially when initialized', () => {
-    // Arrange
-    const props = {
-      workspace,
-      storageDetails: _.merge(defaultGoogleBucketOptions, defaultAzureStorageOptions),
-    };
-
-    // Act
-    render(h(BucketLocation, props));
-
-    // Assert
-    expect(screen.queryByText('Loading')).not.toBeNull();
-  });
-
-  it('renders the bucket location if available, and has no accessibility errors', async () => {
-    // Arrange
-    const props = {
-      workspace,
-      storageDetails: _.mergeAll([
-        defaultGoogleBucketOptions,
-        { fetchedGoogleBucketLocation: 'SUCCESS' },
-        defaultAzureStorageOptions,
-      ]),
-    };
-
-    // Act
-    const { container } = render(h(BucketLocation, props));
-
-    // Assert
-    expect(screen.queryByText('Loading')).toBeNull();
-    expect(screen.getAllByText(/Iowa/)).not.toBeNull();
-    expect(await axe(container)).toHaveNoViolations();
-  });
-
-  it('fetches the bucket location if workspaceContainer attempt encountered an error', async () => {
-    // Arrange
-    const props = {
-      workspace,
-      storageDetails: _.mergeAll([
-        defaultGoogleBucketOptions,
-        { fetchedGoogleBucketLocation: 'ERROR' },
-        defaultAzureStorageOptions,
-      ]),
-    };
-    asMockedFn(Ajax).mockImplementation(
-      () =>
-        ({
-          Workspaces: {
-            workspace: () =>
-              ({
-                checkBucketLocation: jest.fn().mockResolvedValue({
-                  location: 'bermuda',
-                  locationType: 'triangle',
-                }),
-              } as Partial<AjaxContract['Workspaces']['workspace']>),
-          } as Partial<AjaxContract['Workspaces']>,
-        } as Partial<AjaxContract> as AjaxContract)
-    );
-
-    // Act
-    await act(async () => { render(h(BucketLocation, props)) }) //eslint-disable-line
-
-    // Assert
-    expect(screen.queryByText('Loading')).toBeNull();
-    expect(screen.getAllByText(/bermuda/)).not.toBeNull();
-  });
-
-  it('handles requester pays error', async () => {
-    // Arrange
-    const props = {
-      workspace,
-      storageDetails: _.mergeAll([
-        defaultGoogleBucketOptions,
-        { fetchedGoogleBucketLocation: 'ERROR' },
-        defaultAzureStorageOptions,
-      ]),
-    };
-    const requesterPaysError = { message: 'Requester pays bucket', requesterPaysError: true };
-    asMockedFn(Ajax).mockImplementation(
-      () =>
-        ({
-          Workspaces: {
-            workspace: () =>
-              ({
-                checkBucketLocation: () => Promise.reject(requesterPaysError),
-              } as Partial<AjaxContract['Workspaces']['workspace']>),
-          } as Partial<AjaxContract['Workspaces']>,
-        } as Partial<AjaxContract> as AjaxContract)
-    );
-    // Act
-    await act(async () => { render(h(BucketLocation, props)) }) //eslint-disable-line
-
-    // Assert
-    expect(screen.queryByText('Loading')).toBeNull();
-    expect(screen.getAllByText(/bucket is requester pays/)).not.toBeNull();
-  });
-});
+import { defaultAzureStorageOptions, defaultGoogleBucketOptions } from 'src/testing/workspace-fixtures';
 
 describe('AzureDetails', () => {
   const azureContext = {
