+++ src/pages/workspaces/workspace/Dashboard/WorkspaceInformation.test.ts
@@ -1,0 +1,102 @@
+import { screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import { axe } from 'jest-axe';
+import { h } from 'react-hyperscript-helpers';
+import { azureRegions } from 'src/libs/azure-regions';
+import { WorkspaceInformation } from 'src/pages/workspaces/workspace/Dashboard/WorkspaceInformation';
+import { renderWithAppContexts as render } from 'src/testing/test-utils';
+import {
+  defaultGoogleWorkspace,
+  protectedAzureWorkspace,
+  regionRestrictedAzureWorkspace,
+} from 'src/testing/workspace-fixtures';
+
+jest.mock('src/libs/ajax');
+
+jest.mock('src/libs/notifications');
+
+describe('WorkspaceInformation', () => {
+  it('renders information for a non-protected workspace without region constraints and does not fail accessibility tests', async () => {
+    // Act
+    const { container } = render(
+      h(WorkspaceInformation, { workspace: { ...defaultGoogleWorkspace, workspaceInitialized: true } })
+    );
+
+    // Assert
+    // Access Level
+    expect(screen.getAllByText('Owner')).not.toBeNull();
+    // Created date
+    expect(screen.getAllByText('2/15/2023')).not.toBeNull();
+    // Last updated date
+    expect(screen.getAllByText('3/15/2023')).not.toBeNull();
+    // Should not have workspace protected entry
+    expect(screen.queryByText('Workspace Protected')).toBeNull();
+    // Should not have region constraint
+    expect(screen.queryByText('Region Constraint')).toBeNull();
+
+    // Accessibility
+    expect(await axe(container)).toHaveNoViolations();
+  });
+
+  it('renders information for a protected workspace and does not fail accessibility tests', async () => {
+    const user = userEvent.setup();
+
+    // Act
+    const { container } = render(
+      h(WorkspaceInformation, { workspace: { ...protectedAzureWorkspace, workspaceInitialized: true } })
+    );
+
+    // Assert
+    // Access Level
+    expect(screen.getAllByText('Owner')).not.toBeNull();
+    // Created date
+    expect(screen.getAllByText('2/15/2023')).not.toBeNull();
+    // Last updated date
+    expect(screen.getAllByText('3/15/2023')).not.toBeNull();
+    // Should show protected workspace information.
+    expect(screen.getAllByText('Workspace Protected')).not.toBeNull();
+    // Should not have region constraint
+    expect(screen.queryByText('Region Constraint')).toBeNull();
+
+    // Act, click on the info button to get tooltip text to render.
+    await user.click(screen.getByLabelText('More info'));
+
+    // Assert
+    expect(screen.getAllByText(/protected or sensitive data/)).not.toBeNull();
+
+    // Accessibility
+    expect(await axe(container)).toHaveNoViolations();
+  });
+
+  it('renders information for a workspace with region constraints and does not fail accessibility tests', async () => {
+    const user = userEvent.setup();
+
+    // Act
+    const { container } = render(
+      h(WorkspaceInformation, { workspace: { ...regionRestrictedAzureWorkspace, workspaceInitialized: true } })
+    );
+
+    // Assert
+    // Access Level
+    expect(screen.getAllByText('Owner')).not.toBeNull();
+    // Created date
+    expect(screen.getAllByText('2/15/2023')).not.toBeNull();
+    // Last updated date
+    expect(screen.getAllByText('3/15/2023')).not.toBeNull();
+    // Should not have workspace protected entry
+    expect(screen.queryByText('Workspace Protected')).toBeNull();
+    // Should show region constraint information.
+    expect(screen.getAllByText('Region Constraint')).not.toBeNull();
+
+    // Act, click on the info button to get tooltip text with region labels to render.
+    await user.click(screen.getByLabelText('More info'));
+
+    // Assert
+    expect(screen.getAllByText(new RegExp(`${azureRegions.eastus.label}`))).not.toBeNull();
+    expect(screen.getAllByText(new RegExp(`${azureRegions.westus2.label}`))).not.toBeNull();
+    expect(screen.getAllByText(/unknownRegion/)).not.toBeNull();
+
+    // Accessibility
+    expect(await axe(container)).toHaveNoViolations();
+  });
+});
